remove `decimal/generic_mut_impl.rs`
